Separate Extension Service Functionality
• The iOS clang linker will give a warning if an iOS app extension uses a framework that uses UIKit in some way
• To fix this (and to improve separation of concerns), we will be splitting out Notification Service Extension functionality into its own separate framework
