+++ iOS_SDK/OneSignalDevApp/OneSignalNotificationServiceExtension/NotificationService.m
@@ -1,9 +1,6 @@
-#import <OneSignal/OneSignal.h>
-
-#import <UIKit/UIKit.h>
 
 #import "NotificationService.h"
-
+#import <OneSignalExtension/OneSignalExtension.h>
 #import <FirebaseAnalytics/FIRApp.h>
 #import <FirebaseAnalytics/FIRAnalytics.h>
 
@@ -62,7 +59,7 @@
     
     NSLog(@"START!!!!!! request.content.userInfo: %@", request.content.userInfo);
     
-    [OneSignal didReceiveNotificationExtensionRequest:self.receivedRequest withMutableNotificationContent:self.bestAttemptContent];
+    [OneSignalExtension didReceiveNotificationExtensionRequest:self.receivedRequest withMutableNotificationContent:self.bestAttemptContent];
     // DEBUGGING: Uncomment the 2 lines below and comment out the one above to ensure this extension is excuting
     //            Note, this extension only runs when mutable-content is set
     //            Setting an attachment or action buttons automatically adds this
@@ -90,7 +87,7 @@
     // Called just before the extension will be terminated by the system.
     // Use this as an opportunity to deliver your "best attempt" at modified content, otherwise the original push payload will be used.
     
-    [OneSignal serviceExtensionTimeWillExpireRequest:self.receivedRequest withMutableNotificationContent:self.bestAttemptContent];
+    [OneSignalExtension serviceExtensionTimeWillExpireRequest:self.receivedRequest withMutableNotificationContent:self.bestAttemptContent];
     
     self.contentHandler(self.bestAttemptContent);
 }
