+++ iOS_SDK/OneSignalSDK/OneSignalExtension/NSURLSession+OneSignal.h
@@ -1,0 +1,29 @@
+//
+//  NSURLSession+OneSignal.h
+//  OneSignal
+//
+//  Created by Brad Hesse on 9/26/18.
+//  Copyright © 2018 Hiptic. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface DirectDownloadDelegate : NSObject <NSURLSessionDataDelegate> {
+    NSError* error;
+    NSURLResponse* response;
+    BOOL done;
+    NSFileHandle* outputHandle;
+}
+@property (readonly, getter=isDone) BOOL done;
+@property (readonly) NSError* error;
+@property (readonly) NSURLResponse* response;
+
+@end
+
+NS_ASSUME_NONNULL_END
+
+@interface NSURLSession (DirectDownload)
++ (NSString *)downloadItemAtURL:(NSURL *)url toFile:(NSString *)localPath error:(NSError **)error;
+@end
