+++ iOS_SDK/OneSignalSDK/OneSignalExtension/NSURLSession+OneSignal.m
@@ -1,0 +1,77 @@
+//
+//  NSURLSession+OneSignal.m
+//  OneSignal
+//
+//  Created by Brad Hesse on 9/26/18.
+//  Copyright © 2018 Hiptic. All rights reserved.
+//
+
+#import "NSURLSession+OneSignal.h"
+
+@implementation DirectDownloadDelegate
+@synthesize error, response, done;
+
+-(void)URLSession:(NSURLSession *)session dataTask:(NSURLSessionDataTask *)dataTask didReceiveData:(NSData *)data {
+    [outputHandle writeData:data];
+}
+
+-(void)URLSession:(NSURLSession *)session dataTask:(NSURLSessionDataTask *)dataTask didReceiveResponse:(NSURLResponse *)aResponse completionHandler:(void (^)(NSURLSessionResponseDisposition))completionHandler {
+    response = aResponse;
+    completionHandler(NSURLSessionResponseAllow);
+}
+
+-(void)URLSession:(NSURLSession *)session didBecomeInvalidWithError:(NSError *)anError {
+    error = anError;
+    done = YES;
+    
+    [outputHandle closeFile];
+}
+
+-(void)URLSession:(NSURLSession *)session task:(NSURLSessionTask *)task didCompleteWithError:(NSError *)anError {
+    done = YES;
+    error = anError;
+    [outputHandle closeFile];
+}
+
+- (id)initWithFilePath:(NSString*)path {
+    if (self = [super init]) {
+        if ([[NSFileManager defaultManager] fileExistsAtPath:path])
+            [[NSFileManager defaultManager] removeItemAtPath:path error:nil];
+        
+        [[NSFileManager defaultManager] createFileAtPath:path contents:nil attributes:nil];
+        outputHandle = [NSFileHandle fileHandleForWritingAtPath:path];
+    }
+    return self;
+}
+@end
+
+@implementation NSURLSession (DirectDownload)
+
++ (NSString * _Nullable)downloadItemAtURL:(NSURL * _Nonnull)url toFile:(NSString * _Nonnull)localPath error:(NSError **)error {
+    NSMutableURLRequest *request = [[NSMutableURLRequest alloc] initWithURL:url];
+    
+    DirectDownloadDelegate *delegate = [[DirectDownloadDelegate alloc] initWithFilePath:localPath];
+    
+    NSURLSession *session = [NSURLSession sessionWithConfiguration:[NSURLSessionConfiguration defaultSessionConfiguration] delegate:delegate delegateQueue:nil];
+    
+    NSURLSessionDataTask *task = [session dataTaskWithRequest:request];
+    
+    [task resume];
+    
+    [session finishTasksAndInvalidate];
+    
+    while (![delegate isDone]) {
+        [[NSRunLoop currentRunLoop] runUntilDate:[NSDate dateWithTimeIntervalSinceNow:1.0]];
+    }
+    
+    NSError *downloadError = [delegate error];
+    if (downloadError != nil) {
+        if (error)
+            *error = downloadError;
+        return nil;
+    }
+    
+    return delegate.response.MIMEType;
+}
+
+@end
