+++ iOS_SDK/OneSignalSDK/Source/NSURL+OneSignal.h
@@ -31,4 +31,6 @@
 
 - (NSString *)valueFromQueryParameter:(NSString *)parameter;
 
+- (BOOL)isWWWScheme;
+
 @end
