+++ iOS_SDK/OneSignalSDK/Source/OSNotificationPayload.m
@@ -29,8 +29,6 @@
 
 #import "OSNotificationPayload+Internal.h"
 
-#import "OneSignal.h"
-
 @implementation OSNotificationPayload
 
 +(instancetype)parseWithApns:(NSDictionary*)message {
